Image and Invoice file uplaods and downloads

+++ controllers/admin.js
@@ -1,6 +1,10 @@
 const Product = require("../models/product")
+
 const { validationResult } = require("express-validator")
+
 const mongoose = require("mongoose")
+
+const fileHelper = require("../util/file")
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -141,7 +145,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId
   const updatedTitle = req.body.title
   const updatedPrice = req.body.price
-  const updatedImageUrl = req.body.imageUrl
+  const image = req.file
   const updatedDesc = req.body.description
 
   const errors = validationResult(req)
@@ -153,7 +157,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId,
@@ -171,7 +174,10 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle
       product.price = updatedPrice
       product.description = updatedDesc
-      product.imageUrl = updatedImageUrl
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl)
+        product.imageUrl = image.path
+      }
       return product.save().then((result) => {
         console.log("UPDATED PRODUCT!")
         res.redirect("/admin/products")
@@ -212,14 +218,22 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then(() => {
-      console.log("DESTROYED THE PRODUCT")
-      res.redirect("/admin/products")
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product not found."))
+      }
+      fileHelper.deleteFile(product.imageUrl)
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
+        .then(() => {
+          console.log("DESTROYED THE PRODUCT")
+          res.redirect("/admin/products")
+        })
+        .catch((err) => {
+          const error = new Error(err)
+          error.httpStatusCode = 500
+          return next(error)
+        })
     })
-    .catch((err) => {
-      const error = new Error(err)
-      error.httpStatusCode = 500
-      return next(error)
-    })
+    .catch((err) => next(err))
 }
